+++ at_follows_flutter/lib/screens/connections.dart
@@ -2,6 +2,7 @@ import 'package:at_follows_flutter/domain/connection_model.dart';
 import 'package:at_follows_flutter/services/connections_service.dart';
 import 'package:at_follows_flutter/services/sdk_service.dart';
 import 'package:at_follows_flutter/services/size_config.dart';
+import 'package:at_follows_flutter/utils/app_constants.dart';
 import 'package:at_follows_flutter/utils/color_constants.dart';
 import 'package:at_follows_flutter/utils/custom_textstyles.dart';
 import 'package:at_follows_flutter/utils/strings.dart';
@@ -146,7 +147,11 @@ class _ConnectionsState extends State<Connections> {
                       TextField(
                         style: CustomTextStyles.fontR16primary,
                         onChanged: (value) {
-                          setState(() {});
+                          if (value == '') {
+                            FocusScope.of(context).unfocus();
+                          } else {
+                            setState(() {});
+                          }
                         },
                         textInputAction: TextInputAction.search,
                         controller: searchController,