+++ at_follows_flutter/test/unit_tests/connections_service_test.dart
@@ -176,9 +176,11 @@ void main() {
         ..key = AppConstants.following
         ..metadata = metadata;
       await _sdkService.put(atKey1, '@sameeraja🛠,@sitaram🛠');
+      AppConstants.appNamespace = 'sample';
 
       var followersValue = await _sdkService
           .scanAndGet('${AppConstants.followers}|${AppConstants.followersKey}');
+      print('followers value is $followersValue');
       expect(followersValue.value.isNotEmpty, true);
 
       var oldKeyfollowersValue =
@@ -187,6 +189,8 @@ void main() {
 
       var followingValue = await _sdkService
           .scanAndGet('${AppConstants.following}|${AppConstants.followingKey}');
+      print('following value is $followingValue');
+
       expect(followingValue.value.isNotEmpty, true);
 
       var oldKeyfollowingValue =